factor out concepts_work_item.hpp
