+++ include/tmc/spawn_func.hpp
@@ -7,11 +7,11 @@
 
 #include "tmc/detail/compat.hpp"
 #include "tmc/detail/concepts.hpp" // IWYU pragma: keep
+#include "tmc/detail/concepts_work_item.hpp"
 #include "tmc/detail/mixins.hpp"
 #include "tmc/detail/thread_locals.hpp"
 #include "tmc/detail/unsafe_task.hpp"
 #include "tmc/ex_any.hpp"
-#include "tmc/task.hpp"
 #include "tmc/work_item.hpp"
 
 #include <cassert>
